requirements.txt
asgiref==3.5.2,
Django==4.1.2,
djangorestframework==3.14.0,
future==0.18.2,
Pillow==9.3.0,
psycopg2==2.9.5,
pytz==2022.5,
sqlparse==0.4.3,
tzdata==2022.5,
Django~=4.1.2
djangorestframework~=3.14.0